+++ js/custom-setups/card-setup.js
@@ -1,12 +1,9 @@
 import { minMaxes } from "./min-max.js";
 
 export function customCardSetup(s) {
-    if (_.startsWith(s.id, "soccer") || _.startsWith(s.id, "indoor-lacrosse")) {
-        customWidthHeightCardSetup(
-            s.id,
-            s.appearance.width,
-            s.appearance.height
-        );
+    customWidthHeightCardSetup(s.id, s.appearance.width, s.appearance.height);
+    if (s.id === "ice-hockey-iihf") {
+        customCornerRadiusSetup(s.id, s.appearance.cornerRadius);
     }
 }
 
@@ -25,10 +22,13 @@ function customWidthHeightCardSetup(id, width, height) {
         .attr("name", `${id}-width`)
         .attr("min", minMax.minWidth)
         .attr("max", minMax.maxWidth)
-        .attr("value", width);
-    widthField
-        .append("span")
-        .text(`(min: ${minMax.minWidth}, max: ${minMax.maxWidth})`);
+        .attr("value", width)
+        .attr("disabled", minMax.minWidth === minMax.maxWidth ? true : null);
+    if (minMax.minWidth !== minMax.maxWidth) {
+        widthField
+            .append("span")
+            .text(`(min: ${minMax.minWidth}, max: ${minMax.maxWidth})`);
+    }
 
     const heightField = dim.append("div");
     heightField.append("label").attr("for", `${id}-height`).text("Height:");
@@ -53,6 +53,45 @@ function customWidthHeightCardSetup(id, width, height) {
         let params = new URLSearchParams({
             width: width,
             height: height,
+        });
+        window.location.href = `./${id}?${params.toString()}`;
+    });
+}
+
+function customCornerRadiusSetup(id, cornerRadius) {
+    const minMax = minMaxes[id];
+    const card = d3.select(`#${id}`).attr("href", undefined);
+    const dim = card.select(".card-text").select(".dimensions");
+    const cornerRadiusField = dim.append("div");
+    cornerRadiusField
+        .append("label")
+        .attr("for", `${id}-corner-radius`)
+        .text("Corner Radius:");
+    cornerRadiusField
+        .append("input")
+        .attr("id", `${id}-corner-radius`)
+        .attr("type", "number")
+        .attr("name", `${id}-corner-radius`)
+        .attr("min", minMax.minCornerRadius)
+        .attr("max", minMax.maxCornerRadius)
+        .attr("step", 0.1)
+        .attr("value", cornerRadius);
+    cornerRadiusField
+        .append("span")
+        .text(
+            `(min: ${minMax.minCornerRadius}, max: ${minMax.maxCornerRadius})`
+        );
+
+    card.select("button").on("click", function () {
+        const width = d3.select(`#${id}-width`).property("value");
+        const height = d3.select(`#${id}-height`).property("value");
+        const cornerRadius = d3
+            .select(`#${id}-corner-radius`)
+            .property("value");
+        let params = new URLSearchParams({
+            width: width,
+            height: height,
+            cornerRadius: cornerRadius,
         });
         window.location.href = `./${id}?${params.toString()}`;
     });